Reduce Gateway request logging diff
